include Peierls phase factors in measurements
For now, only when complex numbers are used.
Previously the (complex) code outputted incorrect current correlators,
kinetic energy correlators, etc due to missing these factors. Now the
measurements g00, gt0, kv, kn, kk, ksks, jj, jsjs include these factors.
These measurements are gauge invariant.

notes when using a perpendicular field:
0) The non-interacting Hamiltonian is like
sum_{i j} -t_{i j} P_{i j} c_i^dagger c_j
P_{i j} = exp(i 2 pi/Phi_0 int_i^j A(r) dot dr)
where Phi_0 is the flux quantum.
The line integral from i to j depends on the path. Here we use the
shortest straight-line path.

PhysRevB.65.115104 section IIIB is a nice reference.

The total flux through the torus must be an integer multiple of Phi_0.
This defines the parameter nflux. The field is
vec{B} = B nflux*Phi_0/N \hat{z},
where N is the number of unit cells (total area if each unit cell has
area 1).
Choose
A(r) = -alpha B y \hat{x} + (1-alpha) B x \hat{y}.
For example alpha = 0.5 gives the symmetric gauge.

1) g00 and gt0 are now defined as <c_i c_j^\dagger P_{j i}>
where the Peierls phase P_{j i} is defined as above.
In the Hamiltonian, i and j are always close to each other so there is a
clear choice for the path between i and j. In Green's functions, i and j
can be anything. If i_x - j_x is exactly Nx/2 (similarly for the y
direction), it is not clear which path to take. In these cases we choose
arbitrarily to perform the line integral with a path that goes left (or
down). See line 176 in util/gen_1band_hub.py.

This way, g00 and gt0 are gauge invariant but path dependant
measurements.

g00 and gt0 should not be fourier transformed to g(k) or anything
that's a function of k because translation symmetry is completely broken
by choosing a gauge. Very interesting discussion about this may be found
in the Appendix of PhysRevB.82.134510

2) density, double_occ, nn (density-density), zz (spin-z spin-z),
nem_nnnn, nem_ssss
measurements are always gauge invariant

3) kv, kn, kk, vv, vn, ksks, jj, jsjs measurements are always gauge
invariant because they are made from gauge-invariant pieces of the
Hamiltonian.

4) xx (spin-x spin-x) measurements are gauge invariant if the Peierls
phases are the same for spin-up and spin-down. They are not gauge
invariant if the phases are opposite for spin-up and spin-down. It is
possible to make them gauge invariant by multiplying by two factors of
P_ij, similar to what's done in g00 and gt0. Not currently done in the
code.

5) (For the attractive Hubbard model, the Holstein model, and similar, we
would want the Peierls phase to be opposite for spin-up and spin-down to
keep things sign-free. For the half-filled repulsive Hubbard model, we
want the same phase for spin-up and spin-down)
pair_sw (local s-wave singlet) measurements are gauge invariant if the
Peierls phase is opposite for spin-up and spin-down. If the phase is the
same, it is possible to make them gauge invariant by multiplying by two
factors of P_ij, similar to what's done in g00 and gt0. Not currently
done in the code.

6) pair_bb (bond-singlet) measurements are not gauge-invariant in either
case. I don't know what the best thing to do is.